Fix unseen count
