+++ src/modules/notifications/selectors/notifications.js
@@ -1,8 +1,11 @@
 import { createSelector } from "reselect";
+
 import { selectNotificationsState } from "src/select-state";
+import store from "src/store";
 
 import * as notificationLevels from "modules/notifications/constants/notifications";
 
+import { augur } from "services/augurjs";
 import getValue from "utils/get-value";
 
 export const selectNotificationsByLevel = level => state =>
@@ -18,7 +21,18 @@ export const selectInfoNotifications = selectNotificationsByLevel(
 export const selectInfoNotificationsAndSeenCount = createSelector(
   selectInfoNotifications,
   notifications => {
+    const networkId = augur.rpc.getNetworkID();
+    const { universe } = store.getState();
+
     const sortedNotifications = notifications
+      .filter(
+        notification =>
+          (notification.networkId === networkId.toString() &&
+            notification.universe === universe.id) ||
+          typeof notification.networkId === "undefined" ||
+          typeof notification.universe === "undefined"
+      )
+      .reverse()
       .map((notification, i) => ({
         ...notification,
         index: i
